requirements.txt
python-dotenv>=1.0.0
numpy>=1.26.0
uagents>=0.22.0
requests>=2.31.0
pillow>=10.0.0
opencv-python-headless>=4.10.0
qrcode[pil]>=7.4.2
reportlab>=4.0.0
pybraille>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
